pyproject.toml
[project]
name = "cinema-django"
version = "0.1.0"
description = "web development project on Django for Samir"
requires-python = ">=3.10"
dependencies = [
    "django>=5.2.12",
    "django-cors-headers>=4.9.0",
    "django-filter>=25.2",
    "django-stubs>=6.0.1",
    "djangorestframework>=3.16.1",
    "djangorestframework-stubs>=3.16.8",
    "fastapi>=0.136.1",
    "httpx>=0.28.1",
    "pytest>=8.4.2",
    "pytest-cov>=4.1",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.46.0",
]

[tool.pytest.ini_options]
addopts = [
    "--cov=src",
    "--cov-branch",
    "--cov-report=term-missing",
    "-q",
]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.coverage.run]
source = ["src"]
branch = true
parallel = true
omit = ["*/__init__.py",]

[tool.coverage.report]
show_missing = true
fail_under = 80
